+++ src/components/App/Pages/Admin/Admin.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Admin() {
+    return (
+        <nav>
+            <h1 className='admin'>ADMIN PAGE</h1>
+        </nav>
+    )
+}
