+++ src/components/App/Pages/Home/Home.css
@@ -1,0 +1,14 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  font-family: "Montserrat", sans-serif;
+}
+
+.webpage {
+  margin: 2.5rem;
+}
+
+.Welcome {
+  text-align: center;
+}
