+++ src/components/App/Pages/Home/counter.jsx
