+++ src/components/App/Pages/Mentee/Mentee.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Mentee() {
+    return (
+        <nav>
+            <h1 className='mentee'>MENTEE PAGE</h1>
+        </nav>
+    )
+}
