+++ src/components/App/Pages/Mentor/Mentor.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Mentor() {
+    return (
+        <nav>
+            <h1 className='mentor'>MENTOR PAGE</h1>
+        </nav>
+    )
+}
