+++ src/components/App/Pages/SignUp/SignUp.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SignUp() {
+    return (
+        <nav>
+            <h1 className='signup'>SIGN UP PAGE</h1>
+        </nav>
+    )
+}
