+++ src/components/NavBar/Button.js
@@ -1,30 +1,11 @@
 import React from 'react';
 import './Button.css';
+import { Link } from 'react-router-dom';
 
-const STYLES = [
-    'btn--primary', // primary colour of button
-    'btn--outline'  
-]
-
-const SIZES = [
-    'btn--medium',  // for button size
-    'btn--large'
-]
-
-export const Button = ({
-    children,
-    type,
-    onClick,
-    buttonStyle,
-    buttonSize
-}) => {
-    const checkButtonStyle = STYLES.includes(buttonStyle) ? buttonStyle : STYLES[0]
-
-    const checkButtonSize = SIZES.includes(buttonSize) ? buttonSize : SIZES[0]
-
+export function Button() {
     return (
-        <button className={`btn ${checkButtonStyle} ${checkButtonSize}`} onClick={onClick} type={type}>
-            {children}
-        </button>
+        <Link to='sign-up'>
+            <button className='btn'>Sign Up</button>
+        </Link>
     )
 }
