+++ src/components/NavBar/NavBar.css
@@ -11,18 +11,39 @@
   color: #fff;
   justify-self: start;
   margin-left: 20px;
-  cursor: pointer;
+  font-size: 2.4rem;
 }
 
 .nav-menu {
   display: grid;
-  grid-template-columns: repeat(5, auto);
+  grid-template-columns: repeat(3, auto);
   column-gap: 10px;
   list-style: none;
   text-align: center;
   width: 70vw;
   justify-content: end;
   margin-right: 2rem;
+  margin-top: 1rem;
+}
+
+.sign-up {
+  list-style: none;
+  margin-right: 2rem;
+}
+
+.sign_up {
+  color: #6d76f7;
+  text-decoration: none;
+  padding: 8px 20px;
+  border-radius: 4px;
+  background: #fff;
+}
+
+.sign_up:hover {
+  color: white;
+  background-color: #9196e4;
+  border-radius: 4px;
+  transition: all 0.6s ease-out;
 }
 
 .nav-links {
@@ -55,13 +76,17 @@
     position: relative;
   }
 
+  .navbar-logo {
+    font-size: 1.7rem;
+  }
+
   .nav-menu {
     display: flex;
     flex-direction: column;
     width: 100%;
-    height: 500px;
+    height: 276px;
     position: absolute;
-    top: 80px;
+    top: 63px;
     left: -100%;
     opacity: 1;
     transition: all 0.5s ease;
@@ -108,19 +133,9 @@
     font-size: 2rem;
   }
 
-  .nav-links-mobile {
-    display: block;
-    text-align: center;
-    padding: 2rem;
-    width: 100%;
-    display: table;
-    color: white;
-    text-decoration: none;
-  }
-
-  .nav-links-mobile:hover {
-    background-color: #7577fa;
-    border-radius: 0;
-    transition: all 0.6s ease-out;
+  .sign_up {
+    position: absolute;
+    left: 70%;
+    top: 17px;
   }
 }
